requirements.txt
# Core API Dependencies
google-genai>=0.7.0
openai>=1.0.0
anthropic>=0.20.0

# Google Services
google-api-python-client>=2.100.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0

# Media Processing
ffmpeg-python>=0.2.0
Pillow>=10.0.0
mutagen>=1.47.0

# Data Processing
pydantic>=2.0.0
requests>=2.31.0

# System Utilities
psutil>=5.9.0

# Development (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
